refactor!: Refactor built-in scalars to enum
Refactors the `BUILT_IN_TYPE` object to an enum for easier and safer
access to "all built-in types". Also, renames it to `BuiltInScalars`
because it only contains scalars.

This does not attempt to solve any issues with built-in scalars yet
but serve as starting point for later refactorings.

Also fixes two `IllegalArgumentExceptions` that should have become
`InvalidInputValueException`s with #65.