feat: Create dedicated Introspection class
This creates a dedicated `Introspection` class to hold functionality
regarding schema introspection.

As added benefit, we now provide introspection queries for (currently)
two different levels of the GraphQL specification.

This is supposed to be a starting point for future refactoring, as
a) there are still parts left in other places of the code, and b) the
way how that introspection query is composed leaves... room for
improvement (graphql-java for example has a configurable builder).