feat: Export schema in SDL
Adds support for exporting an existing schema in the human-readable
schema definition language (SDL). The Ktor feature by default provides
it via `GET /<endpoint>?schema`.

SDL follows introspection config, so if introspection queries are
disabled, export in SDL will also be disabled.

As implementations that support SDL have to provide the `@deprecated`
directive, this also adds that directive to the built-in directives
(without attempting to fix existing issues there).

Lastly, this fixes a small inconsistency with the spec by making the
`description` of a `__Type` optional (and subsequently gets rid of
some artificial empty string defaults).

Resolves #79