refactor: Reduce coupling with Type
The current implementation is heavily coupled to `Type` which itself
is heavily coupled to actual Kotlin classes. This tries to move away
from `Type` when it is not needed to simplify working with non-Kotlin
schemas (like schemas created from SDL or introspection queries).

Also makes `SchemaProxy` a `__Schema` only to reduce the necessity for
unused overrides and removes various unused properties and functions.