feat!: Automatic input type naming
To avoid equal names between input and query types, input types now
get an automated `"Input"` suffix unless their name already ends with
`"Input"` or they have explicit configuration in the DSL via
`inputType {}`. This avoids in a lot of cases an otherwise unnecessary
manual configuration step.

BREAKING CHANGE: This may change existing schemas when a certain type
was only used as input type and is not configured explicitly (technically
also when a type was used as input _and_ query type but that resulted
in an invalid schema anyway).