refactor!: remove DataLoaderPreparedRequestExecutor
Adds proper handling for `dataProperties` to `ParallelRequestExecutor`
and gets rid of `DataLoaderPreparedRequestExecutor`, along with the
option to configure the request executor in `SchemaConfiguration`.

This also consolidates some tests that were hardcoded to use one of
the request executors, and therefore missed the other.

Resolves #58
Resolves #268

BREAKING CHANGE: The option to configure request executors has been
removed, and execution will now always use the `ParallelRequestExecutor`.