docs: enhance documentation (#323)
Adds documentation for #322 along with a general overhaul of existing
examples in order to make the documentation a bit less cluttered and
easier to read.

Examples are now consistently shown as tabbed, with multiple tabs for
code, request, response, and SDL where appropriate. Some existing
examples have been extended and/or fixed to compile.