feat!: support dataProperty with ParallelRequestExecutor
Adds proper handling for `dataProperties` to `ParallelRequestExecutor`
and gets rid of `DataLoaderPreparedRequestExecutor`, along with the
option to configure the request executor in `SchemaConfiguration` and
related `DataLoaderPreparedRequestExecutor`-specific options.

This also consolidates some tests that were hardcoded to use one of
the request executors, and therefore missed the other, and extends
coverage for dataProperties a bit. Further, this reduces visibility of
some functions to `internal` to reduce namespace pollution.

Resolves #58
Resolves #268

BREAKING CHANGE: The option to configure request executors has been
removed, and execution will now always use the `ParallelRequestExecutor`.
BREAKING CHANGE: The option to configure a timeout has been removed,
as it was only respected by `DataLoaderPreparedRequestExecutor`.
BREAKING CHANGE: The possibility to pass `ExecutionOptions` to schema
execution has been removed, as it was only respected by `DataLoaderPreparedRequestExecutor`.
BREAKING CHANGE: The `DataLoaderPreparedRequestExecutor`, `Executor`,
and `ExecutionOptions` classes have been removed.